+++ gradle/libs.versions.toml
@@ -8,8 +8,6 @@ quilted_fabric_api = "2.0.0-alpha.2+0.55.3-1.19"
 
 kdl = "0.2.0"
 
-static_data = "1.1.2"
-
 [libraries]
 minecraft = { module = "com.mojang:minecraft", version.ref = "minecraft" }
 quilt_mappings = { module = "org.quiltmc:quilt-mappings", version.ref = "quilt_mappings" }
@@ -19,8 +17,6 @@ quilted_fabric_api = { module = "org.quiltmc.quilted-fabric-api:quilted-fabric-a
 quilted_fabric_api_deprecated = { module = "org.quiltmc.quilted-fabric-api:quilted-fabric-api-deprecated", version.ref = "quilted_fabric_api" }
 
 kdl = { module = "dev.hbeck.kdl:kdl4j", version.ref = "kdl" }
-
-static_data = { module = "io.github.cottonmc:StaticData", version.ref = "static_data" }
 
 # If you have multiple similar dependencies, you can declare a dependency bundle and reference it on the build script with "libs.bundles.example".
 [bundles]
