+++ src/main/resources/quilt.mod.json
@@ -33,10 +33,6 @@
 			{
 				"id": "minecraft",
 				"versions": ">=1.18.2"
-			},
-			{
-				"id": "staticdata",
-				"versions": "*"
 			}
 		]
 	}
